Make time field display understand user-supplied widths
The editor doesn't yet understand these, and so it will overflow, but
I'll eventually fix that.

+++ src/ui/view/layout.rs
@@ -46,5 +46,5 @@ pub struct Index {
     /// The index being displayed.
     pub index: crate::model::time::position::Index,
     /// The number of digits to display for this index.
-    pub num_digits: u8,
+    pub num_digits: usize,
 }
